Add files via upload
These files are the scan images with the black markings indicating particles I got coordinates from to use in Freud. All scans are 2 x 2 um. All scans show GaAs TSQDs grown at 535 C, with 4.5 ML of GaAs being deposited onto the sample surface. In terms of the naming on each file, the "T" at the end of certain files were used to indicate images that I had taken, to differentiate from the scans that Chris took before me. These images were used in imagej to get the x and y coordinates. Note that there is one scan that was misplaced and is being searched for.